requirements.txt
# AI Glasses Coach Backend Dependencies

# FastAPI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# AI/ML APIs
anthropic>=0.39.0
openai>=1.0.0

# Math/Science processing
sympy>=1.12
scipy>=1.11.0
numpy>=1.24.0

# Caching
redis>=5.0.0

# Database
aiosqlite>=0.19.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
pillow>=10.0.0

# HTTP client for async requests
httpx>=0.25.0

# Gemini AI (new unified SDK)
google-genai>=1.0.0
websockets>=12.0

# LiveKit Voice Agent
livekit-agents[google]>=1.3.0
livekit-api>=1.0.0

# Voice Tools
yfinance>=0.2.0
caldav>=1.3.0
python-dateutil>=2.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
respx>=0.20.0
